level popup overlapping properly, style modularization (not yet complete)
