+++ client/components/Game/Game.jsx
@@ -3,20 +3,23 @@ import Board from './Board.jsx';
 import NavBar from '../NavBar.jsx';
 import Level from '../Level.jsx';
 
-const Game = () => {
-  const [selectedLevel, setSelectedLevel] = useState('intermediate');
-  const [showLevelOptions, setShowLevelOptions] = useState(false);
+const Game = ({ selectedLevel }) => {
+  // const [selectedLevel, setSelectedLevel] = useState('intermediate');
+  // const [showLevelOptions, setShowLevelOptions] = useState(false);
 
-  const toggleLevelPopup = () => {
-    setShowLevelOptions(!showLevelOptions);
-  };
+  // const toggleLevelPopup = () => {
+  //   setShowLevelOptions(!showLevelOptions);
+  // };
 
   return (
     <div id="game">
-      <NavBar onLevelButtonClick={toggleLevelPopup} />
-      {showLevelOptions && <Level selectedLevel={selectedLevel} onLevelChange={setSelectedLevel} />}
+      {/* <NavBar onLevelButtonClick={toggleLevelPopup} />
+      {showLevelOptions && <Level selectedLevel={selectedLevel} onLevelChange={setSelectedLevel} />} */}
       <h2>Codesweeper</h2>
-      <Board selectedLevel={selectedLevel} onLevelChange={setSelectedLevel} />
+      <Board 
+        selectedLevel={selectedLevel} 
+        // onLevelChange={setSelectedLevel} 
+      />
     </div>
   );
 };
