+++ client/index.js
@@ -2,6 +2,6 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App.jsx';
 
-import styles from './scss/app.scss'
+// import styles from './scss/app.scss'
 
 ReactDOM.render(<App />, document.getElementById('root'));
