+++ client/scss/board.scss
@@ -1,0 +1,91 @@
+@import 'color_palette';
+#game {
+  display: flex;
+  position: relative;
+  flex-direction: column;
+  justify-content: center;
+  // flex-grow: 1;
+  // // margin: 5vh;
+  // align-self: center;
+  // height: 100vmin;
+  // // height: 100%;
+  // width: 100vmin;
+  // // min-height: 94vmin;
+  z-index: 0;
+};
+
+#stats {
+  display: flex;
+  justify-content: space-around;
+  align-content: center;
+  align-self: center;
+  width: 100%;
+  font-size: 5vmin;
+  height: 10vmin;
+  margin-bottom: 2vmin;
+  background-color: $darkblue;
+  box-shadow: inset 2px 2px 2px 0 darken($darkblue, 20), inset -2px -2px 2px 0 lighten($darkblue, 20);
+  z-index: 0;
+
+
+  #timer {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    color: white;
+    width: 40%;
+    margin: 1%;
+    box-shadow: inset 2px 2px 2px 0 lighten($blue, 20), inset -2px -2px 2px 0 darken($blue, 20);
+    background-color: darken($grey-blue, 15);
+    border-radius: 5px;
+  };
+
+  #smile {
+    background-color: darken($blue, 20);
+    font-size: 4.5vmin;
+    text-align: center;
+    border-radius: 5px;
+    box-shadow: inset 2px 2px 2px 0 lighten($blue, 20), inset -2px -2px 2px 0 darken($blue, 20);
+
+    cursor: pointer;
+    margin: 1%;
+    width: 20%;
+  };
+
+  #mineCount {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    // justify-self: start; //orig
+    font-family: 'Orbitron', sans-serif;
+    color: white;
+    // font-size: 6vmin;
+    // text-align: center;
+    width: 40%;
+    margin: 1%;
+    box-shadow: inset 2px 2px 2px 0 lighten($blue, 20), inset -2px -2px 2px 0 darken($blue, 20);
+    background-color: darken($grey-blue, 15);
+    border-radius: 5px;
+    z-index: 0;
+
+  };
+};
+
+#board{
+  display: flex;
+  flex-direction: column;
+  justify-items: center;
+  align-items: center;
+  width: 62vmin;
+  height: 75vmin;
+  align-self: center;
+  z-index: 0;
+};
+
+#grid {
+  display: grid; 
+  align-self: center;
+  width: 100%;
+  height: 100%;
+  z-index: 0;
+};
