+++ client/scss/global.scss
@@ -1,0 +1,35 @@
+@import url('https://fonts.googleapis.com/css2?family=Orbitron:wght@400;500;600;700&display=swap');
+
+@import 'color_palette';
+
+html, body {
+  top: 0;
+  margin: 0;
+  padding: 0;
+};
+
+body {
+  background: darken(rgb(56, 97, 140), 25);
+  font-family: 'Orbitron', sans-serif;
+};
+
+h2 {
+  text-align: center;
+  font-size: 8vmin;
+  color: white;
+  margin: 3.5vh;
+  margin-top: 0;
+  z-index: 0;
+};
+
+#app {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  margin: 0;
+  padding: 0;
+  top: 0;
+  width: 100%;
+  height: 100%;
+};
