+++ client/scss/nav-bar.scss
@@ -1,0 +1,31 @@
+@import 'color_palette';
+
+
+.navBar {
+  display: flex;
+  margin: 0;
+  align-items: center;
+  justify-content: space-between;
+  top: 0;
+  height: 5vh;
+  width: 100%;
+  font-family: 'Orbitron', sans-serif;
+  z-index: 999;
+};
+
+.navLink {
+  font-family: 'Orbitron', sans-serif;
+  font-size: 2.5vh;
+  color: $green;
+  background-color: darken($grey-blue, 25);
+  box-shadow: none;
+  display: flex;
+  justify-content: space-around;
+  align-items: center;
+  height: 100%;
+  width: 25%;
+  padding: 10px;
+  margin: 5%;
+  border-radius: 5px;
+  z-index: 999;
+};
